requirements.txt
# Django core
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.15.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

mysqlclient==2.2.4

# PostgreSQL Driver
psycopg2-binary>=2.9.9

# CORS (nếu frontend React sẽ gọi API)
django-cors-headers>=4.3.1

# Image processing
Pillow>=10.0.0

# CSV export and data manipulation
pandas>=2.2.2

# Environment variables
python-dotenv>=1.0.1
